Refactor: Rename 360-Degree Image Viewer component to ThreeSixtyDegreeImageViewer and update related imports in stories for consistency. Introduce new component implementation with rotation and zoom functionality.
